hexen: Make -playdemo cope with paths.
Vanilla Hexen makes you specify the demo name to play by giving
the plain lump name, eg. heretic -playdemo mydemo to load mydemo.lmp.
It doesn't work if you specify the extension or the full file path.
As a convenience and to match the behavior of Chocolate Doom, allow
paths and extensions.

Also rework the code for other games so that they're slightly more
consistent. This fixes #301.