music: Suggest using .ogg format, not .mp3.
Using the .mp3 extension in generated music config files carries an
implicit recommendation to use that format. There are a number of
reasons to prefer other formats. The biggest is that MP3 is patent
encumbered. Furthermore, while FLAC and Ogg Vorbis both support the
Vorbis comment format that allows metadata loop point tags to be
specified.