opl: Handle negative time division values.
A negative time division file indicates the MIDI file uses SMPTE time
rather than the normal time system. This is not supported yet, but for
the time being, return a sensible time division value that doesn't
cause the sound to stutter and the game to become unplayable.

This fixes #352, although the affected MIDIs do not yet play properly.