Add desktop entries for all the games, make the Setup launch generic
The same icon is used for all four games, even though it's designed to
resemble Doom's logo and looks a bit silly with all of them lined up.

Also the Setup Desktop Entry now only launches chocolate-setup, which
then prompts for the game to configure. This avoids cluttering up a
DE's Preferences menu, control panel, or however it will be displayed.