Fix alpha channel for icons.
Tweak the convert-icon script to include the alpha channel value
as well as RGB. Fix i_video and the setup code to set the icon
including the alpha channel.

While we're at it, replace the current PNG icons with 128x128 high
quality versions so that they look nice on retina display Macs.