doom: Change weaponowned[] to an int array.
The st_stuff.c status bar code passes pointer to an entry in
weaponowned[] as an int pointer, but weaponowned[] is an array
of booleans and booleans are not always represented as 32-bit ints.
Remove the (int *) cast and ensure correct behavior.

This fixes the immediate issue in #509 (thanks floppes).