textscreen: Remove screen size-based font selection.
This was removed on Windows a while back in favour of a more
conservative approach that examined the system DPI settings, but the
logic still remained for other systems and it really isn't a good
idea. Remove the logic that chooses a large font on large monitors,
and add a TODO comment to add proper Linux font selection in the
future, based on the Gtk+ HiDPI setting.