doom: the sky changes in final2 map11->12 and map20->21 transitions.
The id Anthology version of doom2.exe with Final Doom fixed the "Sky
never changes in Doom II" bug. The original Doom source release
included the repaired code, but it was re-bugged in an early Chocolate
Doom version to emulate the behavior of the better-known versions of
vanilla.

Thanks to @fabiangreffrath for demonstrating how he (more-generically)
fixed the bug in Crispy Doom.

Closes #533