update I_StartSound, S_StartSound with pitch
Re-introduce removed code for pitch shifting. In the case of Doom,
bring back the code from the linux doom sources. In the case of
Heretic and Hexen, uncomment existing code.

The original implementation passed a pitch value around through
several of the sound routines, including I_StartSound and
I_UpdateSoundParams, likely because DMX required it. We won't
need it in I_UpdateSoundParams, so I haven't brought those bits
back.

Adjust the sound driver prototypes for the pitch variable.