Replace ->driver_data with GetAllocatedSoundBySfxinfo
GetAllocatedSoundBySfxinfo is a linked-list look-up on the
list of allocated sounds, rather than the direct-access of the
driver_data pointer.

Previously we used the ->driver_data field of sfxinfo_t to point at the
corresponding allocated_sound_t, but this assumes a 1:1 mapping which
won't be true with pitch-shifted sounds.

The lookups are still cheap, but I did some instrumenting to see
how many iterations this results in. Note that when I ran these tests
there were a few more calls that I've since eliminated, so this is
slightly worse than current. For Doom 2 v1.9 demos:

        ____demo:___
 iter     1   2    3
    0   296 636 1691
    1   104 205 620
    2   57  134 488
    3   55  90  287
    4   58  56  204
    5   28  48  146
    6   18  33  119
    7   6   44  99
    8   4   33  86
    9   8   57  72
    10  6   30  62
    11  6   18  68
    12  6   22  46
    13  2   18  42
    14  4   18  48
    15  0   6   22
    16  4   12  16
    17  0   12  14
    18  0   8   12
    19  0   12  14
    20  0   16  12
    21  0   10  10
    22  0   2   20
    23  0   10  10
    24  0   2   6
    25  0   2   8
    26  0   0   10
    27  0   0   4
    28  0   0   6
    29  0   0   2
    30  0   0   12
    31  0   0   8
    32  0   0   4
    33  0   0   2
    34  0   0   4
    35  0   0   0
    36  0   0   0
    37  0   0   0