Enable pitch-shifting in Hexen
Hexen modified the sfxinfo_t structure from Doom, removing the
pitch member (which was used for sfx links, which aren't used
in Hexen anyway) and adding 'changePitch', a flag which indicated
whether a given sound effect should be shifted, or not.

Rather than add the member back, overload the 'pitch' member to
mean the same thing.

The table of sfxinfo_t definitions was already writing the
changePitch value into the pitch member, so no further changes
are needed there.