textscreen: Add TXT_SetTableColumns() function.
It's common to want to divide a window into columns, but this has
until now only been possible by adding an inner table with multiple
columns. To simplify laying out windows, this function allows the
number of columns in a window (or any table) to be changed after it
is allocated.