osx: Add CFBundleIdentifier to Info.plist.
The preferences for the OS X launcher are currently stored in a file
ambiguously named "launcher.plist" that could conflict with any other
app on the system that chose to use the same moniker. Use the more
specific identifier of org.chocolate-doom.launcher to avoid conflicts.

Also set NSHumanReadableCopyright so that we get more detail on the
"About" window in the launcher app.

This fixes #570. Thanks to Xeriphas1994 for the suggestion to tweak
the identifier.