opl: Add a secret DMXOPTION to reverse stereo channels.
The OPL3 support in Doom's DMX library includes stereo panning support,
but there is a bug where the stereo channels are backwards for some
reason. By default preserve this behavior so that we behave like
Vanilla, but add a secret DMXOPTION so that the channels can be
"reversed" into the correct positions by people who really want it.

Thanks to CSonicGo (@sneakernets) for the original pull request to fix
this (#586), and Alexey Khokholov for assistance in diagnosing the bug.