wad: Free cached lumps on reload, don't release.
When reloading a PWAD we discard any lumps that were cached from
its previous incarnation. But it is important that we completely
free the cached lumps and eliminate them from the zone heap,
otherwise the zone memory system might try to free cached lumps
later on and mistakenly overwrite the owner pointer that is no
longer valid.